Move speechQueue to TTS file
